requirements.txt
matplotlib
plotly
pandas	# required to plot images with plotly
pyyaml
aim
numpy
# for learner
torch
# for verification
z3-solver
dreal
sympy
# for convex-opt controller
cvxpy
cvxpylayers
# for rl algorithms
gymnasium[mujoco]
stable-baselines3
tensorboard
tyro
moviepy
pygame